+++ service-user/src/main/java/com/example/microservices/serviceuser/dao/mapper/UserMapper.java
@@ -45,11 +45,11 @@ public interface UserMapper {
      * @return com.example.microservices.serviceuser.dao.entity.User
      * @throws
      */
-    User getUser(Integer userId);
+    User getUser(Long userId);
 
     void insertUser(User user);
 
     void updateUser(User user);
 
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 }
